Set up tests showing captions in both live and VOD segments.
We're having some troubles with some devices not being able to read
captions data from the stream, so the two test statements and the
segments added here go to show that the captions data does exist.

Run the tests via the following command:
    npm run build-test && npm run test:node

A test expectation is set up and also the captions contents are
logged to the console.